% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gaussian.R
\name{fit_gaussian}
\alias{fit_gaussian}
\title{Generic Fit Gaussian Function}
\usage{
fit_gaussian(x, y, s = 30, m = NA)
}
\arguments{
\item{x}{}

\item{y}{}

\item{s}{}

\item{m}{}
}
\description{
Generic Fit Gaussian Function
}

