% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfapply.R
\name{lfapply}
\alias{lfapply}
\title{Apply a Function over a long-form dataset column by variable.}
\usage{
lfapply(df, col_expr, col_n, fn = mean)
}
\arguments{
\item{df}{}

\item{col_expr}{}

\item{col_n}{}

\item{fn}{}
}
\description{
Apply a Function over a long-form dataset column by variable.
}
\examples{
data <- SPPdispersion

avg <- lfapply(data, reflection, angle, mean)
se <- lfapply(data, reflection, angle, sd) / sqrt(20)
ymin <- avg - se
ymax <- avg + se

result <- tibble(step = 1:225, avg, se, ymin, ymax)
ggplot(result, aes(step, avg, ymin = ymin, ymax = ymax)) + geom_line() +
  geom_errorbar()+theme_bw()
}

