% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.timedynamics.R
\name{read.timedynamics}
\alias{read.timedynamics}
\title{Time Dynamics Plot}
\usage{
read.timedynamics(filename, av, time_max = 4, time_offset = 0,
  zero_offset = 1:10, measurement_type = "\%", lock_in_mV = 100,
  quick_plot = T, ...)
}
\arguments{
\item{filename}{}

\item{av}{}

\item{time_max}{}

\item{time_offset}{}

\item{zero_offset}{}

\item{measurement_type}{}

\item{lock_in_mV}{}

\item{quick_plot}{}

\item{...}{}
}
\description{
Plot from time dynamics kit dataset, providing mean and standard error for a given number of averages
}
\examples{

}
\author{
Tom Constant
}

