\name{SPPdispersion}
\alias{SPPdispersion}
\docType{data}
\title{
Example Wavelength-Angle-Reflectivity Dataset
}
\description{
Example Wavelength-Angle-Reflectivity Dataset, in this case for a SPP excited an an asymmetrix zigzag grating.
}
\usage{data(SPPdispersion)}
\format{
  A data frame with 11925 observations on the following 3 variables.
  \describe{
    \item{\code{wavelength}}{wavelength in m}
    \item{\code{angle}}{angle in degrees}
    \item{\code{reflection}}{reflectivity}
  }
}
\details{
Surface plasmon dispersion data from (unpublished) work by author.
}
\source{
Author
}
\references{
None
}
\examples{
data(SPPdispersion)
## maybe str(SPPdispersion) ; plot(SPPdispersion) ...
}
\keyword{datasets}
