\name{ultrafastMonkey-package}
\alias{ultrafastMonkey-package}
\alias{ultrafastMonkey}
\docType{package}
\title{
Misc Tools for work in Exeter University's Ultrafast Laser Lab
}
\description{
Misc Tools for work in Exeter University's Ultrafast Laser Lab. Including shoetcuts for plotting, reading formatted data and quick plots. 
}
\details{
\tabular{ll}{
Package: \tab ultrafastMonkey\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-18\cr
License: \tab What license is it under?\cr
}

}
\author{
Tom Constant

Maintainer: Who to complain to <tc340@ex.ac.uk>

}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
